refactor(tests): keep test schemas out of TestApp
This keeps the unit test graphql pretty isolated.
TestApp now takes an option parameter.
