+++ tests/TestApp.js
@@ -15,9 +15,8 @@ import schemas from "../imports/node-app/devserver/schemas";
 import resolvers from "../imports/node-app/devserver/resolvers";
 
 class TestApp {
-  constructor() {
+  constructor(options = { extraSchemas: [] }) {
     this.collections = {};
-
     this.app = createApolloServer({
       addCallMeteorMethod(context) {
         context.callMeteorMethod = (name) => {
@@ -33,7 +32,7 @@ class TestApp {
         mutations,
         queries
       },
-      typeDefs: schemas,
+      schemas: [...schemas, ...options.extraSchemas],
       resolvers
       // Uncomment this if you need to debug a test. Otherwise we keep debug mode off to avoid extra
       // error logging in the test output.
